Yield `ERB` block as contents
Invoke the `link_to` helper with the original arguments, save for the
modified `options` Hash.

Prior to this commit, the `current_link_to` helper was unable to yield
a block and [capture] the contents to be rendered as children to the
resulting [`<a>` element][mdn-a].

The presence or absence of the [`&block` as an argument][ruby-block] is
_only_ useful to determine which index in the argument list references
the `link_to` invocation's desired `path` value.

Since the `text` (or contents of the [`yield`-ed][ruby-yield] block) is
unmodified, we can delegate to the underlying `link_to` helper with the
original ["splatted" `*arguments` value][ruby-splat].

[capture]: https://middlemanapp.com/basics/helper-methods/#output-helpers
[mdn-a]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a
[ruby-block]: https://ruby-doc.org/core-2.6/Proc.html#class-Proc-label-Creation
[ruby-yield]: https://ruby-doc.org/core-2.2.0/doc/syntax/calling_methods_rdoc.html#label-Proc+to+Block+Conversion
[ruby-splat]: https://ruby-doc.org/core-2.2.0/doc/syntax/calling_methods_rdoc.html#label-Array+to+Arguments+Conversion